+++ src/components/config/channels/whatsapp/components/tabs/AccountTab.vue
@@ -1,47 +1,36 @@
 <template>
-  <div class="general-tab">
-    <div class="general-tab__content">
-      <div class="general-tab__content__info">
-        <div class="general-tab__content__info__account">
-          <div class="general-tab__content__info__account__title">
+  <div class="account-tab">
+    <div class="account-tab__content">
+      <div class="account-tab__content__info">
+        <div class="account-tab__content__info__account">
+          <div class="account-tab__content__info__account__title">
             {{ $t('WhatsApp.config.account.title') }}
           </div>
-          <div class="general-tab__content__info__account__name">
+          <div class="account-tab__content__info__account__name">
             <unnnic-icon-svg
-              class="general-tab__content__info__account__name__icon"
+              class="account-tab__content__info__account__name__icon"
               icon="single-neutral-2"
               scheme="neutral-cloudy"
               size="sm"
             />
-            <div class="general-tab__content__info__account__name__text">Weni Tecnologia</div>
+            <div class="account-tab__content__info__account__name__text">
+              {{ fieldHandler(phoneNumber.display_name) }}
+            </div>
           </div>
-          <div class="general-tab__content__info__account__business">
-            Weni Tech - FB 881768118574045
-          </div>
-        </div>
-
-        <div class="general-tab__content__info__conversations">
-          <div class="general-tab__content__info__conversations__title">
-            {{ $t('WhatsApp.config.conversations.title') }}
-          </div>
-          <div class="general-tab__content__info__conversations__time">
-            conversations start at (local time):
-          </div>
-          <div class="general-tab__content__info__conversations__limit">
-            Message per day limit: 10k
+          <div class="account-tab__content__info__account__business">
+            {{ wabaInfo.name }}
           </div>
         </div>
       </div>
 
       <div>
         <div
-          v-for="(section, index) in generalSections"
+          v-for="(section, index) in accountSections"
           :key="index"
-          class="general-tab__content__section"
+          class="account-tab__content__section"
         >
-          <div class="general-tab__content__section__title">
-            <StatusIndicator :status="section.status" />
-            <span class="general-tab__content__section__title__name">
+          <div class="account-tab__content__section__title">
+            <span class="account-tab__content__section__title__name">
               {{ $t(`WhatsApp.config.${section.name}.title`) }}
             </span>
           </div>
@@ -49,12 +38,12 @@
           <div
             v-for="(field, i) in section.fields"
             :key="i"
-            class="general-tab__content__section__field"
+            class="account-tab__content__section__field"
           >
-            <div class="general-tab__content__section__field__key">
+            <div class="account-tab__content__section__field__key">
               {{ $t(field.label) }}
             </div>
-            <div class="general-tab__content__section__field__value">
+            <div class="account-tab__content__section__field__value">
               {{ field.value }}
             </div>
           </div>
@@ -62,7 +51,7 @@
       </div>
     </div>
     <unnnic-button
-      class="general-tab__close-button"
+      class="account-tab__close-button"
       type="secondary"
       size="large"
       :text="$t('general.Close')"
@@ -72,14 +61,41 @@
 </template>
 
 <script>
-  import StatusIndicator from '../StatusIndicator.vue';
-
   export default {
-    name: 'GeneralTab',
-    components: { StatusIndicator },
-    data() {
-      return {
-        generalSections: [
+    name: 'AccountTab',
+    props: {
+      appInfo: {
+        type: Object,
+        default: /* istanbul ignore next */ () => {},
+      },
+    },
+    methods: {
+      emitClose() {
+        this.$emit('close');
+      },
+      fieldHandler(field) {
+        return field ?? `-`;
+      },
+    },
+    computed: {
+      phoneNumber() {
+        return this.appInfo?.config?.phone_number ?? {};
+      },
+      wabaInfo() {
+        return this.appInfo?.config?.waba ?? {};
+      },
+      appConfig() {
+        return (
+          this.appInfo?.config ?? {
+            phone_number: {},
+            certificate: null,
+            default_template_language: null,
+            consent_status: null,
+          }
+        );
+      },
+      accountSections() {
+        return [
           {
             name: 'channel',
             status: 'green',
@@ -88,49 +104,31 @@
                 type: 'text',
                 name: 'phone_number',
                 label: 'WhatsApp.config.channel.fields.phone_number',
-                value: '+558299620000',
+                value: this.fieldHandler(this.phoneNumber.display_phone_number),
               },
               {
                 type: 'text',
                 name: 'whatsapp_display_name',
                 label: 'WhatsApp.config.channel.fields.whatsapp_display_name',
-                value: 'Weni',
-              },
-              {
-                type: 'text',
-                name: 'number_registration',
-                label: 'WhatsApp.config.channel.fields.number_registration',
-                value: 'Pin Verification',
+                value: this.fieldHandler(this.phoneNumber.display_name),
               },
               {
                 type: 'text',
                 name: 'default_language_for_templates',
                 label: 'WhatsApp.config.channel.fields.default_language_for_templates',
-                value: 'pt_BR',
-              },
-              {
-                type: 'text',
-                name: 'ivr_existing',
-                label: 'WhatsApp.config.channel.fields.ivr_existing',
-                value: 'IVR existing',
-              },
-              {
-                type: 'text',
-                name: 'ivr_active',
-                label: 'WhatsApp.config.channel.fields.ivr_active',
-                value: 'IVR active',
+                value: this.fieldHandler(this.appConfig.default_template_language),
               },
               {
                 type: 'text',
                 name: 'certificate',
                 label: 'WhatsApp.config.channel.fields.certificate',
-                value: 'N/A',
+                value: this.appConfig.certificate ?? 'N/A',
               },
               {
                 type: 'text',
                 name: 'consent_status',
                 label: 'WhatsApp.config.channel.fields.consent_status',
-                value: 'Approved',
+                value: this.fieldHandler(this.appConfig.consent_status),
               },
             ],
           },
@@ -142,47 +140,42 @@
                 type: 'text',
                 name: 'waba_name',
                 label: 'WhatsApp.config.business_account.fields.waba_name',
-                value: 'Weni Tecnologia - FB 881768118500000',
+                value: this.fieldHandler(this.wabaInfo.name),
               },
               {
                 type: 'text',
                 name: 'message_on_behalf_of',
                 label: 'WhatsApp.config.business_account.fields.message_on_behalf_of',
-                value: 'Weni Tecnologia',
+                value: this.fieldHandler(this.wabaInfo.message_behalf_name),
               },
               {
                 type: 'text',
                 name: 'timezone_id',
                 label: 'WhatsApp.config.business_account.fields.timezone_id',
-                value: 'America/Sao_Paulo',
+                value: this.fieldHandler(this.wabaInfo.timezone),
               },
               {
                 type: 'text',
                 name: 'waba_id',
                 label: 'WhatsApp.config.business_account.fields.waba_id',
-                value: '372347354000000',
+                value: this.fieldHandler(this.wabaInfo.id),
               },
               {
                 type: 'text',
                 name: 'namespace',
                 label: 'WhatsApp.config.business_account.fields.namespace',
-                value: '2ee3daabc_0f8e_0000_ae7c_175b808f916',
+                value: this.fieldHandler(this.wabaInfo.namespace),
               },
             ],
           },
-        ],
-      };
-    },
-    methods: {
-      emitClose() {
-        this.$emit('close');
+        ];
       },
     },
   };
 </script>
 
 <style lang="scss" scoped>
-  .general-tab {
+  .account-tab {
     display: flex;
     flex-direction: column;
     height: 100%;
@@ -192,12 +185,12 @@
       flex-direction: column;
       padding-right: $unnnic-spacing-inline-sm;
       overflow-x: hidden;
+      flex: 1;
 
       &__info {
         display: flex;
         justify-content: space-between;
         flex-wrap: wrap;
-        // margin-bottom: $unnnic-spacing-stack-md;
 
         &__account,
         &__conversations {
@@ -218,7 +211,6 @@
             font-size: $unnnic-font-size-body-gt;
             line-height: $unnnic-font-size-body-gt + $unnnic-line-height-md;
             color: $unnnic-color-neutral-cloudy;
-            margin-bottom: $unnnic-spacing-stack-xs;
 
             &__icon {
               margin-right: $unnnic-inline-nano;
@@ -230,11 +222,11 @@
           }
 
           &__business,
-          &__time,
           &__limit {
             font-size: $unnnic-font-size-body-gt;
             line-height: $unnnic-font-size-body-gt + $unnnic-line-height-md;
             color: $unnnic-color-neutral-cloudy;
+            margin-top: $unnnic-spacing-stack-xs;
           }
         }
 
@@ -248,7 +240,6 @@
         flex-direction: column;
         flex-basis: 100%;
         gap: $unnnic-spacing-inline-sm;
-        // margin-bottom: $unnnic-spacing-stack-sm;
 
         &__title {
           display: flex;