+++ src/store/appType/actions.js
@@ -56,4 +56,16 @@ export default {
   async getConversations(store, { code, appUuid, params }) {
     return await appType.getConversations(code, appUuid, params);
   },
+
+  async fetchWppProfile(store, { code, appUuid }) {
+    return await appType.fetchWppProfile(code, appUuid);
+  },
+
+  async updateWppProfile(store, { code, appUuid, payload }) {
+    return await appType.updateWppProfile(code, appUuid, payload);
+  },
+
+  async deleteWppProfilePhoto(store, { code, appUuid }) {
+    return await appType.deleteWppProfilePhoto(code, appUuid);
+  },
 };