+++ tests/unit/specs/store/appType/actions.spec.js
@@ -14,6 +14,9 @@ jest.mock('@/api/appType', () => {
     deleteApp: jest.fn(),
     getSharedWabas: jest.fn(),
     getConversations: jest.fn(),
+    fetchWppProfile: jest.fn(),
+    updateWppProfile: jest.fn(),
+    deleteWppProfilePhoto: jest.fn(),
   };
 });
 
@@ -201,6 +204,44 @@ describe('store/appType/actions.js', () => {
         data.appUuid,
         data.params,
       );
+    });
+
+    it('should call appType.fetchWppProfile', async () => {
+      expect(appTypeApi.fetchWppProfile).not.toHaveBeenCalled();
+      const data = {
+        code: 'code',
+        appUuid: '123',
+      };
+      await actions.fetchWppProfile({}, data);
+      expect(appTypeApi.fetchWppProfile).toHaveBeenCalledTimes(1);
+      expect(appTypeApi.fetchWppProfile).toHaveBeenCalledWith(data.code, data.appUuid);
+    });
+
+    it('should call appType.updateWppProfile', async () => {
+      expect(appTypeApi.updateWppProfile).not.toHaveBeenCalled();
+      const data = {
+        code: 'code',
+        appUuid: '123',
+        payload: {},
+      };
+      await actions.updateWppProfile({}, data);
+      expect(appTypeApi.updateWppProfile).toHaveBeenCalledTimes(1);
+      expect(appTypeApi.updateWppProfile).toHaveBeenCalledWith(
+        data.code,
+        data.appUuid,
+        data.payload,
+      );
+    });
+
+    it('should call deleteWppProfilePhoto', async () => {
+      expect(appTypeApi.deleteWppProfilePhoto).not.toHaveBeenCalled();
+      const data = {
+        code: 'code',
+        appUuid: '123',
+      };
+      await actions.deleteWppProfilePhoto({}, data);
+      expect(appTypeApi.deleteWppProfilePhoto).toHaveBeenCalledTimes(1);
+      expect(appTypeApi.deleteWppProfilePhoto).toHaveBeenCalledWith(data.code, data.appUuid);
     });
   });
 });