+++ tests/unit/specs/store/appType/whatsapp/__mocks__/store.js
@@ -1,0 +1,9 @@
+const mockedContactInfo = {
+  address: 'address',
+  email: 'email',
+  websites: ['website1', 'website2'],
+};
+
+export default {
+  mockedContactInfo,
+};